structured tests
