+++ tests/Models/BasicResource.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ * @property NestedResource $nested
+ */
+class BasicResource extends Model
+{
+    protected $guarded = [];
+
+    protected $keyType = 'string';
+}
