+++ tests/Resources/BasicJsonApiResource.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Resources;
+
+use TiMacDonald\JsonApi\JsonApiResource;
+
+class BasicJsonApiResource extends JsonApiResource
+{
+    //
+}
